+++ docs/reviews/open/I010_review.md
@@ -1,0 +1,46 @@
+# レビュー: I010 科目管理画面の独立
+
+## 基本情報
+- **レビュー対象**: I010 科目管理画面の独立
+- **関連計画書**: docs/plans/open/I010_plan.md
+- **作成日**: 2026-03-27
+
+---
+
+## レビューチェックリスト
+
+### 機能要件
+- [ ] 左サイドメニューに「科目管理」が表示される（org admin のみ）
+- [ ] `/subject-management` に遷移できる
+- [ ] 科目一覧が表示される
+- [ ] 科目を追加できる
+- [ ] 科目追加後に即時更新される
+- [ ] 問題管理画面のアコーディオンが削除されている
+- [ ] 問題管理の既存機能（フィルタ・問題フォーム）に影響がない
+
+### セキュリティ
+- [ ] `role='user'` の場合、科目管理画面にアクセスできない（UI レベル）
+- [ ] バックエンド API は `IsOrgAdmin` パーミッションで保護済み（変更なし）
+
+### コード品質
+- [ ] 削除された import に残留がない（ビルドエラーなし）
+- [ ] `SubjectManagement.tsx` が既存ページ（UserManagement など）と同様のスタイル・構成
+- [ ] TypeScript エラーなし
+
+### テスト
+- [ ] 手動テスト（MT-01〜MT-09）がすべてパス
+- [ ] `npm test` がパス
+- [ ] `npm run build` がパス
+
+---
+
+## レビュー結果
+
+| 項目 | 結果 | 備考 |
+|------|------|------|
+| 機能要件 | - | 実装後に確認 |
+| セキュリティ | - | 実装後に確認 |
+| コード品質 | - | 実装後に確認 |
+| テスト | - | 実装後に確認 |
+
+**総合**: （実装後に記入）
