+++ docs/tests/open/I010_manual_test.md
@@ -1,0 +1,103 @@
+# 手動テスト: I010 科目管理画面の独立
+
+## 前提
+- ローカル環境が起動していること（`docker compose up` または同等）
+- `role='admin'` のユーザー（組織管理者）でログインできること
+- `role='user'` の一般ユーザーでログインできること
+
+---
+
+## テストケース
+
+### MT-01: org admin — サイドメニューに「科目管理」が表示される
+
+1. `role='admin'` ユーザーでログイン
+2. 左サイドメニューを確認
+
+**期待結果**: 「科目管理」メニュー項目が表示されている（School アイコン付き）
+
+---
+
+### MT-02: org admin — 科目管理画面に遷移できる
+
+1. `role='admin'` ユーザーでログイン
+2. 左メニューの「科目管理」をクリック
+
+**期待結果**:
+- URL が `/subject-management` になる
+- 画面タイトル「科目管理 — {組織名}」が表示される
+- 既存の科目一覧が表示される
+
+---
+
+### MT-03: org admin — 科目を追加できる
+
+1. 科目管理画面に遷移
+2. 「科目名」テキストフィールドに新しい科目名を入力
+3. 「追加」ボタンをクリック（または Enter キー）
+
+**期待結果**:
+- 「科目を追加しました」トーストが表示される
+- 入力フィールドがクリアされる
+- 科目一覧に新しい科目が即時追加される
+
+---
+
+### MT-04: org admin — 空の科目名で追加しようとするとエラー
+
+1. 科目管理画面に遷移
+2. 科目名を空のまま「追加」ボタンをクリック
+
+**期待結果**: 「科目名を入力してください」トーストが表示される
+
+---
+
+### MT-05: 一般ユーザー — サイドメニューに「科目管理」が表示されない
+
+1. `role='user'` ユーザーでログイン
+2. 左サイドメニューを確認
+
+**期待結果**: 「科目管理」メニュー項目が表示されない
+
+---
+
+### MT-06: 一般ユーザー — 直接 URL でアクセスするとアクセス拒否される
+
+1. `role='user'` ユーザーでログイン
+2. ブラウザで `/subject-management` に直接アクセス
+
+**期待結果**: 「このページには組織管理者のみアクセスできます。」という Alert が表示される
+
+---
+
+### MT-07: 問題管理画面からアコーディオンが削除されている
+
+1. `role='admin'` ユーザーでログイン
+2. 「問題管理」に遷移
+
+**期待結果**: 「科目管理 — {組織名}」アコーディオンが表示されない
+
+---
+
+### MT-08: 問題管理画面の既存機能が正常動作する
+
+1. `role='admin'` ユーザーでログイン
+2. 「問題管理」に遷移
+3. 科目フィルタのドロップダウンを確認
+4. 「問題を追加」から新規問題追加ダイアログを開き、科目選択を確認
+
+**期待結果**:
+- 科目フィルタのドロップダウンに科目一覧が表示される
+- 問題追加フォームの「科目」セレクトに科目一覧が表示される
+
+---
+
+### MT-09: モバイル表示でサイドメニューが機能する
+
+1. `role='admin'` ユーザーでログイン
+2. 画面幅を縮小してモバイルレイアウトにする
+3. ハンバーガーメニューを開き「科目管理」をタップ
+
+**期待結果**:
+- ドロワーが閉じる
+- `/subject-management` に遷移する
